+++ src/project/contract/mod.rs
@@ -8,7 +8,7 @@ pub mod metadata;
 pub mod vyper;
 
 use crate::build::contract::Contract as ContractBuild;
-use crate::warning_type::WarningType;
+use crate::message_type::MessageType;
 
 use self::eravm_assembly::Contract as EraVMAssemblyContract;
 use self::llvm_ir::Contract as LLVMIRContract;
@@ -57,7 +57,7 @@ impl Contract {
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         llvm_options: Vec<String>,
         output_assembly: bool,
-        suppressed_warnings: Vec<WarningType>,
+        suppressed_messages: Vec<MessageType>,
         debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<ContractBuild> {
         match self {
@@ -68,7 +68,7 @@ impl Contract {
                 optimizer_settings,
                 llvm_options,
                 output_assembly,
-                suppressed_warnings,
+                suppressed_messages,
                 debug_config,
             ),
             Self::LLVMIR(inner) => inner.compile(
@@ -77,7 +77,7 @@ impl Contract {
                 optimizer_settings,
                 llvm_options,
                 output_assembly,
-                suppressed_warnings,
+                suppressed_messages,
                 debug_config,
             ),
             Self::EraVMAssembly(inner) => inner.compile(
@@ -86,7 +86,7 @@ impl Contract {
                 optimizer_settings,
                 llvm_options,
                 output_assembly,
-                suppressed_warnings,
+                suppressed_messages,
                 debug_config,
             ),
         }